ref: move Exceptions into their own module and inherit from base exception
We were previously inheriting from BaseException -- not good.
